+++ src/CommandPaletteCollect.cpp
@@ -101,6 +101,10 @@ static TempStr UpdateCommandNameTemp(MainWindow* win, int cmdId, Str s) {
             isToggle = true;
             newIsOn = !ShowImageOutlines();
         } break;
+        case CmdToggleTransparencyGrid: {
+            isToggle = true;
+            newIsOn = !ShowTransparencyGrid();
+        } break;
         case CmdToggleLaserPointer: {
             isToggle = true;
             newIsOn = !IsLaserPointerActive();