+++ tests/run-almost-all.ts
@@ -22,6 +22,7 @@ import { testit as issue5840 } from "./issue-5840.ts";
 import { testit as issue5844 } from "./issue-5844.ts";
 import { testit as issue3434 } from "./issue-3434.ts";
 import { testit as cmykImageSave } from "./cmyk-image-save.ts";
+import { testit as issue1809 } from "./issue-1809.ts";
 import { testit as issue5846 } from "./issue-5846.ts";
 import { testit as issue5941 } from "./issue-5941.ts";
 import { testit as issue2447 } from "./issue-2447.ts";
@@ -184,6 +185,7 @@ export const tests: NamedTest[] = [
   ["issue-5844", issue5844],
   ["issue-3434", issue3434],
   ["cmyk-image-save", cmykImageSave],
+  ["issue-1809", issue1809],
   ["issue-5846", issue5846],
   ["issue-5941", issue5941],
   ["issue-2447", issue2447],